some update

+++ subzero.sh
@@ -22,13 +22,16 @@ kill_interval=5
 
 stop_all_async (){
 
+     if [ -z "$1" ];then return 1;fi
+
      cur="$(xdotool getwindowfocus getwindowpid)"
      if ! [ "[$1" == "[$cur" ]; then
 
         sleep $kill_interval
 
        if  ps -p "$1" >/dev/null 2>&1; then
-        echo Stop all async "$1"
+        echo "Stop all async "$1" "$(ps -p "$1" -o comm=)""
+        notify-send "Stop all async "$1" "$(ps -p "$1" -o comm=)""
          kill -STOP "$1" >/dev/null 2>&1
         pgrep -P "$1"  | while read F; do  kill -STOP "$F";done
        fi
@@ -36,6 +39,8 @@ stop_all_async (){
 }
 
 stop_spawned_async (){
+     
+     if [ -z :$1: ];then return 1;fi
 
      cur="$(xdotool getwindowfocus getwindowpid)"
      if ! [ "[$1" == "[$cur" ]; then
@@ -43,7 +48,8 @@ stop_spawned_async (){
         sleep $kill_interval
 
        if  ps -p "$1" >/dev/null 2>&1; then
-        echo Stop spawned async "$1"
+	       echo "Stop spawned async "$1" "$(ps -p "$1" -o comm=)""
+	       notify-send "Stop spawned async "$1" "$(ps -p "$1" -o comm=)""
         pgrep -P "$1"  | while read F; do  kill -STOP "$F";done
        fi
      fi
@@ -56,12 +62,7 @@ while true
      cur="$(xdotool getwindowfocus getwindowpid)"
      if ! [ "[$prev" == "[$cur" ]; then
 
-       echo "
-       ---
-       Unfreezing "$cur"
-       ---
-       "
-
+       echo "Unfreezing "$cur" "$(ps -p "$cur" -o comm=)""
        if  ps -p "$cur" >/dev/null 2>&1; then
          kill -CONT "$cur"  >/dev/null 2>&1 
          pgrep -P "$cur"  | while read F; do  kill -CONT "$F" >/dev/null 2>&1;done
@@ -81,13 +82,8 @@ while true
      cur="$(xdotool getwindowfocus getwindowpid)"
      if ! [ "[$prev" == "[$cur" ]; then
 
-       echo "
-       ---
-       Unfreezing "$cur"
-       ---
-       "
+       echo "Unfreezing "$cur"  "$(ps -p "$cur" -o comm=)""
        pgrep -P "$cur"  | while read F; do kill -CONT "$F";done
-       #pgrep -P "$prev"  | while read F; do  kill -STOP "$F";done
        stop_spawned_async "$prev"
        prev=$cur
      fi
@@ -104,7 +100,7 @@ while true
      cur="$(xdotool getwindowfocus getwindowpid)"
 
      if ! [ "[$prev" == "[$cur" ]; then
-       echo Unfreezing "$cur" 
+       echo "Unfreezing "$cur"  "$(ps -p "$cur" -o comm=)""
        if  ps -p "$cur" >/dev/null 2>&1; then
          kill -CONT "$cur"  >/dev/null 2>&1 
          pgrep -P "$cur"  | while read F; do  kill -CONT "$F" >/dev/null 2>&1;done
